fix(benchmarks): resolve inputset and dtype related issues

+++ benchmarks/x_index_1.py
@@ -26,7 +26,7 @@ def main():
     inputs = []
     labels = []
     for _ in range(100):
-        sample_x = np.random.randint(0, 2 ** 3, size=(3,))
+        sample_x = np.random.randint(0, 2 ** 3, size=(3,), dtype=np.uint8)
 
         inputs.append([sample_x])
         labels.append(function_to_compile(*inputs[-1]))
